requirements.txt
numpy
scipy
matplotlib
pytorch
pyro